add LGPL to config file
the reasoning is that if a dev uses default settings, the GPL doesn't trigger,
and they can just compile the file, since they didn't modify it. If they DO
modify the file, users have the right to know about it, especially since they
might be sneaking in functions or overrides into there that modifies library behavior.

In general, a user is entitled to know how the library has been configured, so that's why
it's being covered by LGPL. That plus I don't want to run a dual license on this project.